Finalized self serv changes in tauri
