+++ web/src-tauri/Cargo.toml
@@ -15,13 +15,13 @@ rust-version = "1.81"
 tauri-build = { version = "2.0.3", features = [] }
 
 [dependencies]
-serde_json = "1.0"
-serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0.133"
+serde = { version = "1.0.216", features = ["derive"] }
 tauri = { version = "2.1.1", features = ["tray-icon"] }
 directories = "5.0.1"
 dirs = "5.0.1"
 # reqwest = { version = "0.12.5", features = ["blocking", "json"] }
-tokio = { version = "1.38.0", features = ["full"] }
+tokio = { version = "1.42.0", features = ["full"] }
 warp = "0.3.7"
 ureq = "=2.12.1"
